backend_py/requirements.txt
confluent-kafka==2.10.1
grpcio==1.73.0
grpcio-tools==1.73.0
protobuf==6.31.1
setuptools==80.9.0
